==> notebooks/1-creating-florida-crosswalks.ipynb <==
# Creating the Florida school number crosswalks

from os import path
import os

import numpy as np
import pandas as pd
from fuzzywuzzy import fuzz
from fuzzywuzzy import process

## Assert folders are in place

folders = [
    '../data/intermediary/keys/',
]

for folder in folders:
    if path.exists(folder):
        print("{folder} is already here!".format(folder=folder))
    else:
        try:
            os.makedirs(folder)
        except OSError:
            print("I couldn't make {folder}!".format(folder=folder))
        else:
            print("{folder} successfully made!".format(folder=folder))

## Broward

keys = pd.read_csv('../data/intermediary/scorecard/names/broward.csv')

raw_attendance = pd.read_csv('../data/input/foia/broward.csv')
attendance = pd.DataFrame()
attendance['school_name'] = raw_attendance['SchoolName'].unique()

choices = keys['school_name_l'].unique()

attendance['key_guess'] = attendance['school_name'].apply(lambda x: process.extract(x, choices, limit=1, scorer=fuzz.token_sort_ratio)[0])
attendance['guess_confidence'] = attendance['key_guess'].apply(lambda x: x[1])
attendance['key_guess'] = attendance['key_guess'].apply(lambda x: x[0])

Check rows that have a low guess confidence score:

attendance[attendance['guess_confidence'] < 75]

Check and see if the amount unique guesses is equal to the number of schools:

print("key guesses:", attendance['key_guess'].nunique())
print("Actual district schools:", attendance['school_name'].nunique())
print("Off by:", attendance['school_name'].nunique() - attendance['key_guess'].nunique())

List out which rows have duplicate guesses:

for school in attendance['key_guess'].unique():
    if (attendance[attendance['key_guess'] == school]['school_name'].nunique()) > 1:
        print("this guess used multiple times:",school)
        print('for these schools:')
        print(attendance[attendance['key_guess'].str.contains(school)]['school_name'].unique())
        print("\n")

List out more potential guesses for those that need fixes:

need_fixes = ['PINE RIDGE ALTERNATIVE CENTER', 'PLANTATION ELEMENTARY']
for need in need_fixes:
    print(need)
    for guess in process.extract(need, choices, limit=10, scorer=fuzz.token_sort_ratio):
        print("\t", guess)

Replace incorrect options:

school = 'PLANTATION ELEMENTARY'
fixed_guess = 'PLANTATION ELEMENTARY SCHOOL'

attendance['key_guess'] = np.where(attendance['school_name'] == school, fixed_guess, attendance['key_guess'])

Remove rows that don't seem to have matching data:

school = 'PINE RIDGE ALTERNATIVE CENTER'
attendance = attendance[attendance['school_name'] != school]

Save as a key file:

KEYS_FILENAME = '../data/intermediary/keys/broward.csv'
df = attendance.merge(keys, left_on='key_guess', right_on='school_name_l')
df[[
    'school_name', 
    'key_guess', 
    'guess_confidence', 
    'district_number', 
    'district_name', 
    'school_number',
    'school_name_l'
]].to_csv(KEYS_FILENAME, index=False)
