==> notebooks/2-cleaning-florida-csv.ipynb <==
# Cleaning csvs

from os import path
import os

import pandas as pd

os.makedirs('../data/output/', exist_ok=True)

## Broward

broward = pd.read_csv('../data/input/foia/broward.csv')
characteristics = pd.read_csv('../data/intermediary/scorecard/characteristics/broward.csv')
equity = pd.read_csv('../data/intermediary/scorecard/equity/broward.csv')
keys = pd.read_csv('../data/intermediary/keys/broward.csv')

broward = broward.rename(columns={'SchoolName': 'school_name'})
broward = broward.merge(keys[['school_name', 'school_number']], on='school_name')

historical = pd.read_csv('../data/input/historical_attendance_1819.csv')
historical = historical[historical['DISTRICT #'] == keys.district_number[0]]
historical.columns

hist_columns = [
    'DAYS PRESENT', 
    'DAYS ABSENT', 
    'TOTAL DAYS', 
    '% PRESENT', 
    'AVERAGE DAILY ATTENDANCE', 
    'AVERAGE DAILY MEMBERSHIP'
]

for column in hist_columns:
    historical[column] = pd.to_numeric(historical[column].str.replace(',','').str.replace('%',''))
    
hist_columns.append("SCHOOL #")

historical_context = pd.merge(historical, equity, left_on="SCHOOL #", right_on="school_number")

historical_poverty = historical_context.groupby('poverty').sum()
historical_poverty['% PRESENT'] = historical_poverty['DAYS PRESENT'] / historical_poverty['TOTAL DAYS']
historical_poverty

export = broward.groupby(['school_name', 'school_number']).sum()
export = export.reset_index()
export['covid.agg.enroll'] = export['Enrollment']
export['covid.agg.attend'] = export['Login']
export = export[['school_name', 'school_number', 'covid.agg.enroll', 'covid.agg.attend']]
export['covid.avg'] = export['covid.agg.attend'] / export['covid.agg.enroll'] * 100

historical = historical.rename(columns={
    "SCHOOL #":      "school_number",
    "DISTRICT #":    "district_number",
    "DISTRICT NAME": "district_name",
    "DAYS PRESENT":  "1819.agg.attend",
    "TOTAL DAYS":    "1819.agg.enroll"
})

historical['1819.avg'] = historical["1819.agg.attend"] / historical["1819.agg.enroll"] * 100

export_hist = historical[[
    "school_number",
    "district_number",
    "district_name",
    "1819.agg.attend",
    "1819.agg.enroll",
    "1819.avg"
]]

export = export.merge(export_hist, on='school_number')

export = export.merge(equity, on=['school_number', 'district_number'])
export = export.merge(characteristics[[
    'school_number',
    'school_type', 
    'grade'
]], on='school_number')

export

export.to_csv("../data/output/broward.csv", index=False)

## Miami-Dade

miami = pd.read_csv('../data/input/foia/miami-dade.csv')
characteristics = pd.read_csv('../data/intermediary/scorecard/characteristics/miami-dade.csv')
equity = pd.read_csv('../data/intermediary/scorecard/equity/miami-dade.csv')
enrollment = pd.read_csv('../data/intermediary/scorecard/enrollment/miami-dade.csv')

df = miami.merge(enrollment, on='school_number')
df['total_students'] = pd.to_numeric(df['total_students'])

df['est_present'] = df['pct_present'] / 100.0 * df['total_students']

historical = pd.read_csv('../data/input/historical_attendance_1819.csv')
historical = historical[historical['DISTRICT NAME'] == 'MIAMI-DADE']
historical.columns

hist_columns = [
    'DAYS PRESENT',
    'TOTAL DAYS', 
    '% PRESENT',
]

for column in hist_columns:
    historical[column] = pd.to_numeric(historical[column].str.replace(',','').str.replace('%',''))
    
hist_columns.append("SCHOOL #")

historical = historical[hist_columns]
historical = historical.rename(columns={
    'TOTAL DAYS':'1819.agg.enroll',
    'DAYS PRESENT':'1819.agg.attend',
    '% PRESENT': '1819.avg',
    'SCHOOL #':'school_number'
})

export = df.groupby(['school_name', 'school_number']).sum()
export = export.reset_index()

export = export[['school_name', 'school_number', 'total_students', 'est_present']]
export = export.merge(characteristics, on="school_number")
export = export.merge(equity, on=["school_number", 'district_number'])

export = export[[
    "school_name",
    "school_number",
    "total_students",
    "est_present",
    "school_type",
    "grade",
    "charter",
    "alt_schl",
    "minority",
    "poverty",
    "title_i"]
]

export = export.merge(historical, on='school_number')

export = export.rename(columns={
    'total_students': 'covid.agg.enroll',
    'est_present': 'covid.agg.attend'
})

export['covid.avg'] = export['covid.agg.attend'] / export['covid.agg.enroll'] * 100

export.to_csv("../data/output/miami-dade.csv", index=False)
