==> notebooks/3-analysis.ipynb <==
# Analysis - Python portion

from collections import OrderedDict

import pandas as pd

## Miami-Dade

The gaps were often wide: some schools in Miami-Dade County dropped to around a 75 percent participation rate for remote learning in April, while many of their peers neared full, 100 percent participation.

miami = pd.read_csv('../data/input/foia/miami-dade.csv')
miami.sort_values(by='pct_present').head(10)

miami[miami.date == '4-6-2020'].sort_values(by='pct_present').tail()

print(len(miami))
print(len(miami[miami.pct_present > 95]))

characteristics = pd.read_csv('../data/intermediary/scorecard/characteristics/miami-dade.csv')
equity = pd.read_csv('../data/intermediary/scorecard/equity/miami-dade.csv')

df = miami.merge(characteristics, on='school_number')
df = df.merge(equity,on='school_number')
df[['school_name', 'date', 'pct_present', 'poverty', 'minority', 'grade']].to_csv('../data/output/miami-dade.csv')

## Broward

broward = pd.read_csv('../data/output/broward.csv')
broward[broward.school_name.str.contains('MCARTHUR') | broward.school_name.str.contains('CYPRESS BAY')]

### Alternative Schools

broward = pd.read_csv('../data/input/foia/broward.csv')
characteristics = pd.read_csv('../data/intermediary/scorecard/characteristics/broward.csv')
equity = pd.read_csv('../data/intermediary/scorecard/equity/broward.csv')
keys = pd.read_csv('../data/intermediary/keys/broward.csv')

broward = broward[['SchoolName', "Enrollment", "Date"]]
broward['Date'] = broward['Date'].astype('datetime64')
broward.set_index(['Date'],inplace=True)

rows = []
for _, d in broward.groupby('SchoolName', sort=False):
    SchoolName = d['SchoolName'][0]
    EnrollmentStart = d['Enrollment'][0]
    EnrollmentEnd = d['Enrollment'][-1]
    row = OrderedDict([
        ('SchoolName' , SchoolName),
        ('EnrollmentStart' , EnrollmentStart),
        ('EnrollmentEnd' , EnrollmentEnd),
        ('EnrollmentDifference', EnrollmentEnd - EnrollmentStart),
        ('EnrollmentChange', (EnrollmentEnd - EnrollmentStart) / EnrollmentStart * 100)
    ])
    rows.append(row)

enrollment = pd.DataFrame(rows)
enrollment = enrollment.rename(columns={'SchoolName':'school_name'})

keys = keys[['school_name', 'school_number']]

analysis = enrollment.merge(keys, on='school_name')
analysis = analysis.merge(equity, on='school_number')
analysis = analysis.merge(characteristics, on='school_number')

# largest decrease in enrollment
analysis.sort_values(by="EnrollmentChange").head(5)

# Only alternative schools
analysis[analysis['alt_schl'] != 'N'].sort_values(by="EnrollmentChange")

## Hillsborough

hillsborough = pd.read_csv('../data/input/foia/hillsborough.csv')
hillsborough['PRESENT'] = hillsborough['ENROLLMENT'] - hillsborough['ABSENT']
hillsborough.sort_values(by="% PRESENT").head(6)

len(hillsborough[hillsborough.ABSENT == 0])

hillsborough[
    (hillsborough['SCHOOL_TYPE'] == 'HIGH') &
    (hillsborough['WEEK'] == "4/10/2020")
].sort_values(by="% PRESENT", ascending=False).head()

characteristics[characteristics['school_number'] == 241]

equity[equity['school_number'] == 241]

## Overall

hillsborough = pd.read_csv('../data/input/foia/hillsborough.csv')
broward = pd.read_csv('../data/input/foia/broward.csv')
miami = pd.read_csv('../data/input/foia/miami-dade.csv')

total_schools = hillsborough.SCHOOLS.nunique() + broward.SchoolName.nunique() + miami.school_name.nunique()
total_schools
